src/software/requirements.txt
fastapi[standard]
psycopg2-binary==2.9.10
cryptography==44.0.0